requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.5.0  # https://github.com/evansd/whitenoise
redis==4.6.0  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
celery==5.3.1  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
flower==2.0.0  # https://github.com/mher/flower
uvicorn[standard]==0.23.1  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==4.2.3  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.3.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.2.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
#drf-spectacular==0.26.3  # https://github.com/tfranzel/drf-spectacular


# ADD
# ------------------------------------------------------------------------------
# dj-rest-auth==2.2.5
djangorestframework-simplejwt
django-filter==23.2
drf-yasg==1.21.5
django-imagekit==4.1.0

# TRANSLATION
django-modeltranslation==0.18.10

django-celery-results==2.5.1

# CHANNELS
Automat==22.10.0
autobahn==23.1.2
channels==4.0.0
constantly==15.1.0
daphne==4.0.0
hyperlink==21.0.0
incremental==22.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyopenssl==23.2.0
service-identity==21.1.0
twisted==22.10.0
txaio==23.1.1
zope.interface==6.0

#CHANNEL_LAYERS REDIS
channels-redis==4.1.0
msgpack==1.0.5

#HTTP/2 SUPPORT
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
priority==1.3.0

# BASE64
drf_base64==2.0

# CKEDITOR
django-ckeditor==6.6.1

# PAYME
paymeuz==0.2
django-payments==2.0.0
drf-spectacular==0.26.5
